+++ library/UnitTestCase.php
@@ -15,10 +15,12 @@ use OxidEsales\Eshop\Core\Module\ModuleVariablesLocator;
 use OxidEsales\Eshop\Core\UtilsObject;
 use OxidEsales\EshopCommunity\Core\Database\Adapter\DatabaseInterface;
 use OxidEsales\EshopCommunity\Core\Database;
+use OxidEsales\TestingLibrary\Helper\ProjectConfigurationHelper;
 use OxidEsales\TestingLibrary\Helper\SessionHelper;
 use OxidEsales\TestingLibrary\Services\Library\DatabaseRestorer\DatabaseRestorerFactory;
 use OxidEsales\TestingLibrary\Services\Library\DatabaseRestorer\DatabaseRestorerInterface;
 
+use OxidEsales\TestingLibrary\Services\Library\ProjectConfigurationHandler;
 use oxTestModules;
 use oxTestsStaticCleaner;
 use PHPUnit\Framework\MockObject\MockBuilder;
@@ -47,7 +49,7 @@ abstract class UnitTestCase extends BaseTestCase
     private static $shopStateBackup;
 
     /** @var VfsStreamWrapper */
-    private static $vfsStreamWrapper;
+    private $vfsStreamWrapper;
 
     /** @var array MultiShop tables used in shop */
     private $multiShopTables = array(
@@ -107,6 +109,7 @@ abstract class UnitTestCase extends BaseTestCase
 
         if ($testConfig->shouldRestoreAfterUnitTests()) {
             $this->backupDatabase();
+            self::getProjectConfigurationHandler()->backup();
         }
 
         \OxidEsales\Eshop\Core\Registry::getUtils()->commitFileCache();
@@ -210,6 +213,7 @@ abstract class UnitTestCase extends BaseTestCase
         if ($testConfig->shouldRestoreAfterUnitTests()) {
             $dbRestore = self::_getDbRestore();
             $dbRestore->restoreDB();
+            self::getProjectConfigurationHandler()->restore();
             \OxidEsales\Eshop\Core\DatabaseProvider::getDb()->closeConnection();
         }
     }
@@ -773,11 +777,11 @@ abstract class UnitTestCase extends BaseTestCase
      */
     public function getVfsStreamWrapper()
     {
-        if (is_null(self::$vfsStreamWrapper)) {
-            self::$vfsStreamWrapper = new VfsStreamWrapper();
+        if ($this->vfsStreamWrapper === null) {
+            $this->vfsStreamWrapper = new VfsStreamWrapper();
         }
 
-        return self::$vfsStreamWrapper;
+        return $this->vfsStreamWrapper;
     }
 
     /**
@@ -1062,5 +1066,13 @@ abstract class UnitTestCase extends BaseTestCase
         if ((PHP_SESSION_ACTIVE == session_status()) && session_id()) {
             session_destroy();
         }
+    }
+
+    /**
+     * @return ProjectConfigurationHandler
+     */
+    private static function getProjectConfigurationHandler(): ProjectConfigurationHandler
+    {
+        return new ProjectConfigurationHandler(new ProjectConfigurationHelper());
     }
 }